backend/requirements.txt
# Core Framework
Django==5.2.3
djangorestframework==3.16

# Security & CORS
django-cors-headers==4.7.0

# Database
psycopg==3.2.9

# HTTP requests
requests==2.32.4

# Environment management
python-dotenv==1.1.1
python-decouple==3.8

# Development tools
ipython==9.3.0

# Google Cloud APIs
google-cloud-vision==3.10.2
google-cloud-storage==3.1.1

# File handling
Pillow==11.2.1

# Code quality
black==25.1.0
flake8==7.3.0
isort==6.0.1
